libathemecore/md5: use size_t instead of int for buffer length

+++ include/md5.h
@@ -50,6 +50,8 @@
  * ARCH_IS_BIG_ENDIAN.
  */
 
+#include <stddef.h> /* size_t */
+
 #ifndef MD5_DIGEST_LENGTH
 #define MD5_DIGEST_LENGTH 16
 #endif
@@ -78,9 +80,9 @@ void md5_init(md5_state_t *pms);
 
 /* Append a string to the message. */
 #ifdef P3
-void md5_append(P3(md5_state_t *pms, const md5_byte_t *data, int nbytes));
+void md5_append(P3(md5_state_t *pms, const md5_byte_t *data, size_t nbytes));
 #else
-void md5_append(md5_state_t *pms, const md5_byte_t *data, int nbytes);
+void md5_append(md5_state_t *pms, const md5_byte_t *data, size_t nbytes);
 #endif
 
 /* Finish the message and return the digest. */
